Make existing note editable. Still trying to figuer out how to save the update back to database properly

+++ src/components/Note.jsx
@@ -1,22 +1,62 @@
-import React from "react";
+import React, { useState } from "react";
 
 function Note(props) {
-  function handleDelete() {
-    props.onDelete(props.id);
+
+  function handleAction() {
+    if (actionButton === "Delete") {
+      props.onDelete(props.id);
+    } else {
+      props.onUpdate(props.id)
+      setActionButton("Delete")
+    }
   }
 
-  function handleEdit() {
-    props.onEdit(props.id);
+  //Know which note is editing -->Checked
+  //Load the content into a state-->Checked
+  //Turn it into editable box-->Checked
+  //Use same method of onchange to manage the content
+  ////Here is a bit of problem, I will need to find a way to pass the edited value back to the state. 
+  //Use the same method of save to save it back to the database. 
+  const [editable, setEditable] = useState("false")
+  const [actionButton, setActionButton] = useState("Delete")
+  const [updateNote, setUpdateNote] = useState({
+    id: "",
+    title: "",
+    content: ""
+  });
+
+  function initEdit() {
+    setUpdateNote({ id: props.id, title: props.title, content: props.content });
+    console.log(updateNote)
+    setEditable("true")
+    setActionButton("Save")
   }
+  function handleChange(event) {
+    //##Error message shows should use event.persist instead of event.currentTarget
+    // console.log(event.currentTarget.textContent)
+    const { name, textContent } = event.target;
+    console.log(name);
+    console.log(event.target.name);
+    console.log(textContent);
+
+    // setUpdateNote(prevNote => {
+    //   return {
+    //     ...prevNote,
+    //     [name]: textContent
+    //   };
+    // });
+    // console.log(updateNote);
+  }
+
 
   return (
-    <div className="note">
-      <button onClick={handleEdit}>Edit</button>
+    <div className="note" >
 
-      <h1>{props.title}</h1>
-      <p>{props.content}</p>
+      <h1 name="title" onClick={initEdit} onInput={handleChange} value={updateNote.content} contentEditable={editable}>{props.title}</h1>
+      <p name="content" onClick={initEdit} onInput={handleChange} value={updateNote.content} contentEditable={editable}>{props.content}</p>
+      <p className="time-mark">Last edited on {props.editOn}</p>
 
-      <button onClick={handleDelete}>DELETE</button>
+      <button onClick={handleAction}>{actionButton}</button>
     </div>
   );
 }
